trusted-types/TrustedTypePolicy-name.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>

<meta http-equiv="Content-Security-Policy" content="trusted-types">
<body>
<script>
  // Policy name test
  test(t => {
    let policy = window.TrustedTypes.createPolicy('hidden', { createHTML: s => s } );
    assert_true(policy instanceof TrustedTypePolicy);
    assert_equals(policy.name, 'hidden');
  }, "policy.name = name");

  // Duplicate names test
  test(t => {
    assert_throws(new TypeError(), _ => {
     window.TrustedTypes.createPolicy('hidden', { createURL: s => s } );
    });
  }, "duplicate policy name attempt throws");

  // Retrieve policy names tests
  // TODO: Change the checks when testharness will be changed to have a policy
  // and to take tests in TrustedHTML
  test(t => {
    let policy = window.TrustedTypes.createPolicy('exposed', { createURL: s => s }, true );
    let names = window.TrustedTypes.getPolicyNames();
    assert_equals(names.length, 2);
    assert_true(names.includes('hidden'));
    assert_true(names.includes('exposed'));
  }, "getPolicyNames returns all policy names");
</script>
